+++ MelonStartScreen/UI/Customization/Config.cs
@@ -23,6 +23,7 @@ namespace MelonLoader.MelonStartScreen.UI.Customization
             internal static MelonPreferences_Entry<Color> Background;
             internal static MelonPreferences_Entry<Color> ProgressBar;
             internal static MelonPreferences_Entry<Color> ProgressBarOutline;
+            internal static MelonPreferences_Entry<Color> Text;
 
             internal static void Setup()
             {
@@ -34,6 +35,7 @@ namespace MelonLoader.MelonStartScreen.UI.Customization
                 Background = Category.CreateEntry(nameof(Background), new Color(0.08f, 0.09f, 0.10f), nameof(Background));
                 ProgressBar = Category.CreateEntry(nameof(ProgressBar), new Color(0.47f, 0.97f, 0.39f), nameof(ProgressBar));
                 ProgressBarOutline = Category.CreateEntry(nameof(ProgressBarOutline), new Color(1.00f, 0.23f, 0.42f), nameof(ProgressBarOutline));
+                Text = Category.CreateEntry(nameof(Text), new Color(1, 1, 1), nameof(Text));
             }
         }
     }