+++ MelonStartScreen/UI/UIStyleValues.cs
@@ -5,25 +5,24 @@ namespace MelonLoader.MelonStartScreen.UI
 {
     internal static class UIStyleValues
     {
-        public static Font standardFont;
+        public static Font TextFont;
 
-        public static Texture2D backgroundTexture;
-        public static Texture2D progressbarInnerTexture;
-        public static Texture2D progressbarOuterTexture;
-
-        public static Texture2D melonloaderLogoTexture;
+        public static Texture2D BackgroundTexture;
+        public static Texture2D ProgressBarTexture;
+        public static Texture2D ProgressBarOutlineTexture;
+        public static Texture2D LogoTexture;
 
         public static AnimatedImage Animation;
 
         internal static void Init()
         {
-            backgroundTexture = UIUtils.CreateColorTexture(Customization.Config.Colors.Background.Value);
-            progressbarInnerTexture = UIUtils.CreateColorTexture(Customization.Config.Colors.ProgressBar.Value);
-            progressbarOuterTexture = UIUtils.CreateColorTexture(Customization.Config.Colors.ProgressBarOutline.Value);
+            BackgroundTexture = UIUtils.CreateColorTexture(Customization.Config.Colors.Background.Value);
+            ProgressBarTexture = UIUtils.CreateColorTexture(Customization.Config.Colors.ProgressBar.Value);
+            ProgressBarOutlineTexture = UIUtils.CreateColorTexture(Customization.Config.Colors.ProgressBarOutline.Value);
 
             MelonDebug.Msg("[UIStyleValues] LoadImage Logo");
-            melonloaderLogoTexture = new Texture2D(2, 2);
-            bool imgLoaded = ImageConversion.LoadImage(melonloaderLogoTexture, Convert.FromBase64String(
+            LogoTexture = new Texture2D(2, 2);
+            bool imgLoaded = ImageConversion.LoadImage(LogoTexture, Convert.FromBase64String(
                 (MelonLaunchOptions.Console.Mode == MelonLaunchOptions.Console.DisplayMode.LEMON)
                 ? ImageDatas.LemonLogo
                 : ImageDatas.MelonLogo), false);
@@ -33,7 +32,7 @@ namespace MelonLoader.MelonStartScreen.UI
             Animation = Customization.Images.LoadAnimatedImage();
 
             // Load Font
-            standardFont = Resources.GetBuiltinResource<Font>("Arial.ttf");
+            TextFont = Resources.GetBuiltinResource<Font>("Arial.ttf");
         }
     }
 }