+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 15. Internal Failures now print their proper Failure Messages in the MessageBox even if not in Debug Mode.
 16. Cleaned Up Start Screen's Animated Image Customization.
 17. Added Start Screen Color Customization for Background, Progress Bar, and Progress Bar Outline.
+18. Added Start Screen Text Color Customization.
 
 ---